разделил код. 5 задание
