+++ index.html
@@ -229,5 +229,6 @@
 
   <script src="js/functions.js"></script>
   <script src="js/main.js"></script>
+  <script type="module" src="main.js"></script>
 </body>
 </html>
